split the base subscription classes
